Update howto_rl_mb_001_robothtm_environment.py following changes in MLPro-SL
